requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
groq
python-dotenv
